+++ trade-analysis/web/src/main/java/com/aura/model/ProvinceTrade.java
@@ -4,8 +4,16 @@ public class ProvinceTrade {
     String provinceName;
     Integer tradeCount;
 
+    public String getProvinceName() {
+        return provinceName;
+    }
+
     public void setProvinceName(String provinceName) {
         this.provinceName = provinceName;
+    }
+
+    public Integer getTradeCount() {
+        return tradeCount;
     }
 
     public void setTradeCount(Integer tradeCount) {
