+++ trade-analysis/web/src/main/resources/c3p0-config.xml
@@ -3,17 +3,17 @@
 
     <default-config>
         <property name="driverClass">com.mysql.jdbc.Driver</property>
-        <property name="jdbcUrl">jdbc:mysql://localhost:3306/aura</property>
+        <property name="jdbcUrl">jdbc:mysql://39.105.196.194:3306/aura</property>
         <property name="user">root</property>
-        <property name="password">root</property>
+        <property name="password">FWfw-1992</property>
         <property name="initialPoolSize">5</property>
         <property name="maxPoolSize">50</property>
     </default-config>
 
     <named-config name="itclass">
         <property name="driverClass">com.mysql.jdbc.Driver</property>
-        <property name="jdbcUrl">jdbc:mysql://localhost:3306/aura</property>
+        <property name="jdbcUrl">jdbc:mysql://39.105.196.194:3306/aura</property>
         <property name="user">root</property>
-        <property name="password">root</property>
+        <property name="password">FWfw-1992</property>
     </named-config>
 </c3p0-config>
