+++ trade-analysis/web/src/main/resources/jdbc.properties
@@ -1,7 +1,7 @@
 driver=com.mysql.jdbc.Driver
-url=jdbc:mysql://localhost:3306/aura?useUnicode=true&characterEncoding=utf-8
+url=jdbc:mysql://39.105.196.194:3306/aura?useUnicode=true&characterEncoding=utf-8
 username=root
-password=root
+password=FWfw-1992
 
 idleConnectionTestPeriodInMinutes=240
 idleMaxAgeInMinutes=60
